minor changes. Also moves the database stuff to its own module
