+++ src/bin/show_users.rs
@@ -1,10 +1,10 @@
 use dotenv::dotenv;
-use test_rocket_app::{init_pool, repositories::users, DbConn};
+use test_rocket_app::{database, repositories::users};
 
 fn main() {
     dotenv().ok();
-    let pool = init_pool();
-    let conn = DbConn(pool.get().unwrap());
+    let pool = database::init_pool();
+    let conn = database::DbConn(pool.get().unwrap());
 
     let results = users::all(&conn).expect("Failed to get users");
     println!("Displaying {} users:", results.len());
